Update Logo and Size (#1829)
This updates the logo to our new logo and sets a much smaller size for link previews. This should help when codecov links are unfurled in Twitter, Slack, etc. 

Note that I sort of guessed on what would work for `og:image:width` to 150px. It looks about right based on my estimation, but I'm not sure what will actually happen until we see it unfurl in slack. Worst case we can just revert, but I think it will be fine. 

Note that I followed this guide to try and determine some changes to make, we could definitely do more here for specific commits, etc. if we really wanted to: https://whitep4nth3r.com/blog/level-up-your-link-previews-in-slack/