One more try at Unfurling
the previous change here lowered the twitter image settings to 600px. I did this to match to default size of the new image (600x600), thinking that Slack honored the `og:...` atrributes for unfurling. Slack does honor these, but it takes preference to twitter attributes if they show up first. So based on this medium post: https://medium.com/slack-developer-blog/everything-you-ever-wanted-to-know-about-unfurling-but-were-afraid-to-ask-or-how-to-make-your-e64b4bb9254

we need to reduce the size for the twitter attributes as well. This will very likely also shrink the thumbnail on twitter, but i'm not sure that's actually a problem.