Attempt Summary card designation
This changes the twitter card type to `summary` from `summary_large_image`. This is my last attempt to shrink the image preview in slack and other locations when app.codecov.io urls are shared. 

docs for this madness are here: https://developer.twitter.com/en/docs/twitter-for-websites/cards/overview/summary

General guidelines state that the large summary card _shouldn't_ be used for application logos, but provides no alternative for what to use if you're representing a logo.