fix: prevent guaranteed toggling of hideFlakeRate between renders
I noticed that any time I would refresh the page or change the ordering of the
table, which would cause useInfiniteTestResults to run, the flake rate column
would disappear for just a second until we got the response back and then the
flake rate column would come back. This fixes this by preventing the value used
in rendering from being updated until the request is done, preventing the
flickering i saw.