feat: add support for all branches in failed tests tab
basically we're going to be encoding a null branch parameter as specifying
"all branches". So if a user goes to `<service>/<owner>/<repo>/tests` they
will no longer get redirected or told to select a branch, we'll first attempt
to fetch the results for all branches from the API.

This change will rely on the change to the API which changes the default value
of the branch when it's undefined to fetch the "all branches" data instead of the
repo's default branch's data.